move diff to base, does not really belong in LinearAlgebra
