+++ stdlib/LinearAlgebra/src/LinearAlgebra.jl
@@ -77,7 +77,6 @@ export
     diag,
     diagind,
     diagm,
-    diff,
     dot,
     eig,
     eigfact,
